NLPerturbator/requirements.txt
datasets==2.7.1
detoxify==0.5.1
ftlid==0.1.2
jellyfish==1.0.1
langid==1.1.6
lemminflect==0.2.2
nltk==3.8.1
numpy==1.21.5
openai==0.27.9
pandas==1.2.4
parrot==1.0
pattern==3.6
PyYAML==6.0.1
sacrebleu==2.3.1
seqeval==1.2.2
spacy==3.4.1
spacytextblob==4.0.0
termcolor==2.3.0
tokenizers==0.13.2
torch==1.13.0
torchtext==0.15.2
tqdm==4.63.1
transformers==4.26.1
